requirements.txt
langchain==0.3.3
pinecone-client==5.0.1
python-dotenv==1.0.1
streamlit==1.39.0